fix integer truncation bug
git-svn-id: http://svn.macosforge.org/repository/ruby/MacRuby/trunk@4526 23306eb0-4c56-4727-a40e-e92c0eb68959

+++ ext/bigdecimal/bigdecimal.c
@@ -2178,7 +2178,7 @@ static int gnAlloc=0; /* Memory allocation counter */
 VP_EXPORT void *
 VpMemAlloc(size_t mb)
 {
-    void *p = xmalloc((unsigned int)mb);
+    void *p = xmalloc(mb);
     if(!p) {
         VpException(VP_EXCEPTION_MEMORY,"failed to allocate memory",1);
     }
